smtp/requirements.txt
aiofiles==23.2.1
aiohttp-socks==0.8.4
aiosignal==1.3.1
aiosmtpd==1.4.5
anyio==4.3.0
async-timeout==4.0.3
atpublic==4.0
attrs==23.2.0
certifi==2024.2.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matrix-nio==0.24.0
multidict==6.0.5
pycryptodome==3.20.0
python-socks==2.4.4
referencing==0.33.0
rpds-py==0.18.0
sniffio==1.3.1
typing_extensions==4.10.0
unpaddedbase64==2.1.0
yarl==1.9.4